refactor: improve project architecture and workflows
