+++ packages/runner/src/execution-cache/run-scenario-cache.ts
@@ -3,7 +3,7 @@ import type {
   RunEvent,
   RunScenarioInput,
   ScenarioRun,
-} from '../execution/run-scenario'
+} from '../execution/run-scenario-types'
 import { requiredValue } from '../required-value'
 import {
   deserializeExecutionCacheEnvelope,
